requirements.txt
streamlit
beautifulsoup4
requests
langchain
langchain-community
chromadb
sentence-transformers
unstructured
python-docx
pdfminer.six
tqdm
ollama
